% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_add.R
\name{add_class_var}
\alias{add_class_var}
\title{Add class variable}
\usage{
add_class_var(
  ds_tb,
  clusters_1L_int,
  model_mdl,
  class_var_nm_1L_chr = "class_int"
)
}
\arguments{
\item{ds_tb}{Dataset (a tibble)}

\item{clusters_1L_int}{Clusters (an integer vector of length one)}

\item{model_mdl}{Model (a model)}

\item{class_var_nm_1L_chr}{Class variable name (a character vector of length one), Default: 'class_int'}
}
\value{
Dataset (a tibble)
}
\description{
add_class_var() is an Add function that updates an object by adding new values to new or empty fields. Specifically, this function implements an algorithm to add class variable. The function returns Dataset (a tibble).
}
