% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_add.R
\name{add_kmean_cls_var}
\alias{add_kmean_cls_var}
\title{Add k-means clustering class variable}
\usage{
add_kmean_cls_var(
  ds_tb,
  classes_1L_int,
  components_1L_int,
  pca_df,
  kmean_var_nm_1L_chr = "kmeans_cls_int",
  start_1L_int = 25L
)
}
\arguments{
\item{ds_tb}{Dataset (a tibble)}

\item{classes_1L_int}{Classes (an integer vector of length one)}

\item{components_1L_int}{Components (an integer vector of length one)}

\item{pca_df}{Principal component analysis (a data.frame)}

\item{kmean_var_nm_1L_chr}{K-means clustering variable name (a character vector of length one), Default: 'kmeans_cls_int'}

\item{start_1L_int}{Start (an integer vector of length one), Default: 25}
}
\value{
Dataset (a tibble)
}
\description{
add_kmean_cls_var() is an Add function that updates an object by adding new values to new or empty fields. Specifically, this function implements an algorithm to add k-means clustering class variable. The function returns Dataset (a tibble).
}
