% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_calculate.R
\name{calculate_mean_RI}
\alias{calculate_mean_RI}
\title{Calculate meanand Index}
\usage{
calculate_mean_RI(
  cvdn_results_ls,
  ds_tb,
  fold_id_nm_1L_chr = "Fold",
  id_var_nm_1L_chr = "ID",
  select_from_ls_1L_int = 2L,
  select_from_df_int = c(1L, 4L)
)
}
\arguments{
\item{cvdn_results_ls}{Cross-validation results (a list)}

\item{ds_tb}{Dataset (a tibble)}

\item{fold_id_nm_1L_chr}{Fold identity name (a character vector of length one), Default: 'Fold'}

\item{id_var_nm_1L_chr}{Identity variable name (a character vector of length one), Default: 'ID'}

\item{select_from_ls_1L_int}{Select from list (an integer vector of length one), Default: 2}

\item{select_from_df_int}{Select from data.frame (an integer vector), Default: c(1L, 4L)}
}
\value{
Meanand Index (a double vector)
}
\description{
calculate_mean_RI() is a Calculate function that performs a numeric calculation. Specifically, this function implements an algorithm to calculate meanand index. The function returns Meanand Index (a double vector).
}
