% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_calculate.R
\name{calculate_sngl_Rand_idx}
\alias{calculate_sngl_Rand_idx}
\title{Calculate singleand index}
\usage{
calculate_sngl_Rand_idx(i, j, data_tb)
}
\arguments{
\item{i}{An object}

\item{j}{An object}

\item{data_tb}{Data (a tibble)}
}
\value{
Rand index (a double vector)
}
\description{
calculate_sngl_Rand_idx() is a Calculate function that performs a numeric calculation. Specifically, this function implements an algorithm to calculate singleand index. The function returns Rand index (a double vector).
}
\keyword{internal}
