% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_calculate.R
\name{calculate_sngl_ri}
\alias{calculate_sngl_ri}
\title{Calculate singleand Index}
\usage{
calculate_sngl_ri(i, j, data_tb)
}
\arguments{
\item{i}{An object}

\item{j}{An object}

\item{data_tb}{Data (a tibble)}
}
\value{
a Rand Index (a double vector)
}
\description{
calculate_sngl_ri() is a Calculate function that performs a numeric calculation. Specifically, this function implements an algorithm to calculate singleand index. The function returns a Rand Index (a double vector).
}
\keyword{internal}
