% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_fit.R
\name{fit_mixture_mdl}
\alias{fit_mixture_mdl}
\title{Fit mixture model}
\usage{
fit_mixture_mdl(ds_tb, clusters_1L_int, var_nms_chr)
}
\arguments{
\item{ds_tb}{Dataset (a tibble)}

\item{clusters_1L_int}{Clusters (an integer vector of length one)}

\item{var_nms_chr}{Variable names (a character vector)}
}
\value{
Model (a model)
}
\description{
fit_mixture_mdl() is a Fit function that fits a model of a specified type to a dataset Specifically, this function implements an algorithm to fit mixture model. The function returns Model (a model).
}
\keyword{internal}
