% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_fit.R
\name{fit_mixture_mdl_clusters}
\alias{fit_mixture_mdl_clusters}
\title{Fit mixture model clusters}
\usage{
fit_mixture_mdl_clusters(ds_tb, nbr_clss_1L_int = 15L, var_nms_chr)
}
\arguments{
\item{ds_tb}{Dataset (a tibble)}

\item{nbr_clss_1L_int}{Number classes (an integer vector of length one), Default: 15}

\item{var_nms_chr}{Variable names (a character vector)}
}
\value{
Model summary (a list)
}
\description{
fit_mixture_mdl_clusters() is a Fit function that fits a model of a specified type to a dataset. Specifically, this function implements an algorithm to fit mixture model clusters. The function returns Model summary (a list).
}
\keyword{internal}
