% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_get.R
\name{get_fake_aqol6d_ds}
\alias{get_fake_aqol6d_ds}
\title{Get fake Assessment of Quality of Life Six Dimension dataset}
\usage{
get_fake_aqol6d_ds()
}
\value{
Dataset (a tibble)
}
\description{
get_fake_aqol6d_ds() is a Get function that retrieves a pre-existing data object from memory, local file system or online repository. Specifically, this function implements an algorithm to get fake assessment of quality of life six dimension dataset. The function returns Dataset (a tibble).
}
