% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_get.R
\name{get_nbr_of_clss}
\alias{get_nbr_of_clss}
\title{Get number of classes}
\usage{
get_nbr_of_clss(cvdn_ds_tb, clss_var_nm_1L_chr = "Classes")
}
\arguments{
\item{cvdn_ds_tb}{Cross-validation dataset (a tibble)}

\item{clss_var_nm_1L_chr}{Classes variable name (a character vector of length one), Default: 'Classes'}
}
\value{
Number of classes (a double vector of length one)
}
\description{
get_nbr_of_clss() is a Get function that extracts data from an object. Specifically, this function implements an algorithm to get number of classes. The function returns Number of classes (a double vector of length one).
}
\keyword{internal}
