% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heterodox-package.R, R/pkg_heterodox.R
\docType{package}
\name{heterodox-package}
\alias{heterodox-package}
\title{heterodox: Explore and Characterise Heterogeneity in Quality of Life Data}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Tools to explore heterogeneity in quality of life data. This development version of the heterodox package has been made available as part of the process of testing and documenting the package. If you have any questions, please contact the authors.

\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Tools to explore heterogeneity in quality of life data. This development version of the heterodox package has been made available as part of the process of testing and documenting the package. If you have any questions, please contact the authors.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://ready4-dev.github.io/heterodox/}
  \item \url{https://github.com/ready4-dev/heterodox}
  \item \url{https://ready4-dev.github.io/ready4/}
}


Useful links:
\itemize{
  \item \url{https://ready4-dev.github.io/heterodox/}
  \item \url{https://github.com/ready4-dev/heterodox}
  \item \url{https://ready4-dev.github.io/ready4/}
}

}
\author{
\strong{Maintainer}: Matthew Hamilton \email{matthew.hamilton1@monash.edu} (\href{https://orcid.org/0000-0001-7407-9194}{ORCID}) [funder, copyright holder]

Authors:
\itemize{
  \item Caroline Gao \email{caroline.gao@orygen.org.au} (\href{https://orcid.org/0000-0002-0987-2759}{ORCID})
}

Other contributors:
\itemize{
  \item Orygen [copyright holder, funder]
  \item Headspace [funder]
  \item National Health and Medical Research Council [funder]
}

}
\keyword{internal}
