% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_make.R
\name{make_cvdn_ds}
\alias{make_cvdn_ds}
\title{Make cross-validation dataset}
\usage{
make_cvdn_ds(
  cvdn_results_ls,
  select_from_ls_1L_int = 1L,
  select_from_df_int = NA_integer_,
  fold_id_nm_1L_chr = "Fold"
)
}
\arguments{
\item{cvdn_results_ls}{Cross-validation results (a list)}

\item{select_from_ls_1L_int}{Select from list (an integer vector of length one), Default: 1}

\item{select_from_df_int}{Select from data.frame (an integer vector), Default: NA}

\item{fold_id_nm_1L_chr}{Fold identity name (a character vector of length one), Default: 'Fold'}
}
\value{
Cross-validation dataset (an output object of multiple potential types)
}
\description{
make_cvdn_ds() is a Make function that creates a new R object. Specifically, this function implements an algorithm to make cross-validation dataset. The function returns Cross-validation dataset (an output object of multiple potential types).
}
