% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_make.R
\name{make_cvdn_ls}
\alias{make_cvdn_ls}
\title{Make cross-validation list}
\usage{
make_cvdn_ls(
  ds_tb,
  folds_int,
  nbr_cores_1L_int = 1L,
  nbr_clss_1L_int = 15L,
  var_nms_chr
)
}
\arguments{
\item{ds_tb}{Dataset (a tibble)}

\item{folds_int}{Folds (an integer vector)}

\item{nbr_cores_1L_int}{Number cores (an integer vector of length one), Default: 1}

\item{nbr_clss_1L_int}{Number classes (an integer vector of length one), Default: 15}

\item{var_nms_chr}{Variable names (a character vector)}
}
\value{
Cross-validation (a list)
}
\description{
make_cvdn_ls() is a Make function that creates a new R object. Specifically, this function implements an algorithm to make cross-validation list. The function returns Cross-validation (a list).
}
