% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_make.R
\name{make_cvdn_points_ds}
\alias{make_cvdn_points_ds}
\title{Make cross-validation points dataset}
\usage{
make_cvdn_points_ds(
  smry_cvdn_ds_tb,
  statistic_var_nm_1L_chr,
  clss_var_nm_1L_chr = "Classes",
  maximise_1L_lgl = F
)
}
\arguments{
\item{smry_cvdn_ds_tb}{Summary cross-validation dataset (a tibble)}

\item{statistic_var_nm_1L_chr}{Statistic variable name (a character vector of length one)}

\item{clss_var_nm_1L_chr}{Classes variable name (a character vector of length one), Default: 'Classes'}

\item{maximise_1L_lgl}{Maximise (a logical vector of length one), Default: F}
}
\value{
Cross-validation points dataset (a data.frame)
}
\description{
make_cvdn_points_ds() is a Make function that creates a new R object. Specifically, this function implements an algorithm to make cross-validation points dataset. The function returns Cross-validation points dataset (a data.frame).
}
\keyword{internal}
