% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_make.R
\name{make_ds_for_ri_calc}
\alias{make_ds_for_ri_calc}
\title{Make dataset forand Index calculation}
\usage{
make_ds_for_ri_calc(cvdn_ds_ls, ds_tb, id_var_nm_1L_chr = "ID")
}
\arguments{
\item{cvdn_ds_ls}{Cross-validation dataset (a list)}

\item{ds_tb}{Dataset (a tibble)}

\item{id_var_nm_1L_chr}{Identity variable name (a character vector of length one), Default: 'ID'}
}
\value{
Transformed dataset (a tibble)
}
\description{
make_ds_for_ri_calc() is a Make function that creates a new R object. Specifically, this function implements an algorithm to make dataset forand index calculation. The function returns Transformed dataset (a tibble).
}
\keyword{internal}
