% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_make.R
\name{make_individual_fit_plt}
\alias{make_individual_fit_plt}
\title{Make individual fit plot}
\usage{
make_individual_fit_plt(
  tfd_cvdn_ds_tb,
  nbr_of_clss_1L_dbl,
  statistic_var_nm_1L_chr,
  clss_var_nm_1L_chr = "Classes",
  fold_var_nm_1L_chr = "Fold",
  maximise_1L_lgl = F,
  y_label_1L_chr = NA_character_
)
}
\arguments{
\item{tfd_cvdn_ds_tb}{Transformed cross-validation dataset (a tibble)}

\item{nbr_of_clss_1L_dbl}{Number of classes (a double vector of length one)}

\item{statistic_var_nm_1L_chr}{Statistic variable name (a character vector of length one)}

\item{clss_var_nm_1L_chr}{Classes variable name (a character vector of length one), Default: 'Classes'}

\item{fold_var_nm_1L_chr}{Fold variable name (a character vector of length one), Default: 'Fold'}

\item{maximise_1L_lgl}{Maximise (a logical vector of length one), Default: F}

\item{y_label_1L_chr}{Y label (a character vector of length one), Default: 'NA'}
}
\value{
Plot (a plot)
}
\description{
make_individual_fit_plt() is a Make function that creates a new R object. Specifically, this function implements an algorithm to make individual fit plot. The function returns Plot (a plot).
}
\keyword{internal}
