% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_make.R
\name{make_pca_tbl}
\alias{make_pca_tbl}
\title{Make principal component analysis table}
\usage{
make_pca_tbl(ds_tb, var_nms_chr, class_var_nm_1L_chr = "class_int")
}
\arguments{
\item{ds_tb}{Dataset (a tibble)}

\item{var_nms_chr}{Variable names (a character vector)}

\item{class_var_nm_1L_chr}{Class variable name (a character vector of length one), Default: 'class_int'}
}
\value{
Principal component analysis (a data.frame)
}
\description{
make_pca_tbl() is a Make function that creates a new R object. Specifically, this function implements an algorithm to make principal component analysis table. The function returns Principal component analysis (a data.frame).
}
