% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_make.R
\name{make_smry_cvdn_ds}
\alias{make_smry_cvdn_ds}
\title{Make summary cross-validation dataset}
\usage{
make_smry_cvdn_ds(
  cvdn_ds_tb,
  statistic_var_nm_1L_chr,
  clss_var_nm_1L_chr = "Classes"
)
}
\arguments{
\item{cvdn_ds_tb}{Cross-validation dataset (a tibble)}

\item{statistic_var_nm_1L_chr}{Statistic variable name (a character vector of length one)}

\item{clss_var_nm_1L_chr}{Classes variable name (a character vector of length one), Default: 'Classes'}
}
\value{
Summary cross-validation dataset (a tibble)
}
\description{
make_smry_cvdn_ds() is a Make function that creates a new R object. Specifically, this function implements an algorithm to make summary cross-validation dataset. The function returns Summary cross-validation dataset (a tibble).
}
\keyword{internal}
