% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_make.R
\name{make_smry_var_nms}
\alias{make_smry_var_nms}
\title{Make summary variable names}
\usage{
make_smry_var_nms(statistic_var_nm_1L_chr, maximise_1L_lgl = F)
}
\arguments{
\item{statistic_var_nm_1L_chr}{Statistic variable name (a character vector of length one)}

\item{maximise_1L_lgl}{Maximise (a logical vector of length one), Default: F}
}
\value{
Summary variable names (a character vector)
}
\description{
make_smry_var_nms() is a Make function that creates a new R object. Specifically, this function implements an algorithm to make summary variable names. The function returns Summary variable names (a character vector).
}
\keyword{internal}
