% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_plot.R
\name{plot_average_fit}
\alias{plot_average_fit}
\title{Plot average fit}
\usage{
plot_average_fit(cvdn_ds_tb, clss_var_nm_1L_chr = "Classes")
}
\arguments{
\item{cvdn_ds_tb}{Cross-validation dataset (a tibble)}

\item{clss_var_nm_1L_chr}{Classes variable name (a character vector of length one), Default: 'Classes'}
}
\value{
Average fit (a plot)
}
\description{
plot_average_fit() is a Plot function that plots data. Specifically, this function implements an algorithm to plot average fit. The function returns Average fit (a plot).
}
