% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_plot.R
\name{plot_cluster}
\alias{plot_cluster}
\title{Plot cluster}
\usage{
plot_cluster(ds_tb, cluster_int)
}
\arguments{
\item{ds_tb}{Dataset (a tibble)}

\item{cluster_int}{Cluster (an integer vector)}
}
\description{
plot_cluster() is a Plot function that plots data Specifically, this function implements an algorithm to plot cluster. The function is called for its side effects and does not return a value.
}
