% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_plot.R
\name{plot_individual_fit}
\alias{plot_individual_fit}
\title{Plot individual fit}
\usage{
plot_individual_fit(
  cvdn_ds_tb,
  clss_var_nm_1L_chr = "Classes",
  fold_var_nm_1L_chr = "Fold",
  legend_label_1L_chr = "Datasets"
)
}
\arguments{
\item{cvdn_ds_tb}{Cross-validation dataset (a tibble)}

\item{clss_var_nm_1L_chr}{Classes variable name (a character vector of length one), Default: 'Classes'}

\item{fold_var_nm_1L_chr}{Fold variable name (a character vector of length one), Default: 'Fold'}

\item{legend_label_1L_chr}{Legend label (a character vector of length one), Default: 'Datasets'}
}
\value{
Individual fit (a plot)
}
\description{
plot_individual_fit() is a Plot function that plots data. Specifically, this function implements an algorithm to plot individual fit. The function returns Individual fit (a plot).
}
