% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_plot.R
\name{plot_pca}
\alias{plot_pca}
\title{Plot principal component analysis}
\usage{
plot_pca(ds_tb, pca_df, class_var_nm_1L_chr = "class_int")
}
\arguments{
\item{ds_tb}{Dataset (a tibble)}

\item{pca_df}{Principal component analysis (a data.frame)}

\item{class_var_nm_1L_chr}{Class variable name (a character vector of length one), Default: 'class_int'}
}
\description{
plot_pca() is a Plot function that plots data Specifically, this function implements an algorithm to plot principal component analysis. The function is called for its side effects and does not return a value.
}
