% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_print.R
\name{print_cluster_plots}
\alias{print_cluster_plots}
\title{Print cluster plots}
\usage{
print_cluster_plots(
  ds_tb,
  clusters_1L_int,
  var_nms_chr,
  nbr_of_folds_1L_int = 2L,
  var_idx_1L_int = 1L
)
}
\arguments{
\item{ds_tb}{Dataset (a tibble)}

\item{clusters_1L_int}{Clusters (an integer vector of length one)}

\item{var_nms_chr}{Variable names (a character vector)}

\item{nbr_of_folds_1L_int}{Number of folds (an integer vector of length one), Default: 2}

\item{var_idx_1L_int}{Variable index (an integer vector of length one), Default: 1}
}
\value{
No return value, called for side effects.
}
\description{
print_cluster_plots() is a Print function that prints output to console. Specifically, this function implements an algorithm to print cluster plots. The function is called for its side effects and does not return a value.
}
