% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_transform.R
\name{transform_cvdn_ds_for_indl_plts}
\alias{transform_cvdn_ds_for_indl_plts}
\title{Transform cross-validation dataset for individual plots}
\usage{
transform_cvdn_ds_for_indl_plts(
  cvdn_ds_tb,
  clss_var_nm_1L_chr = "Classes",
  fold_var_nm_1L_chr = "Fold"
)
}
\arguments{
\item{cvdn_ds_tb}{Cross-validation dataset (a tibble)}

\item{clss_var_nm_1L_chr}{Classes variable name (a character vector of length one), Default: 'Classes'}

\item{fold_var_nm_1L_chr}{Fold variable name (a character vector of length one), Default: 'Fold'}
}
\value{
Transformed cross-validation dataset (a tibble)
}
\description{
transform_cvdn_ds_for_indl_plts() is a Transform function that edits an object in such a way that core object attributes - e.g. shape, dimensions, elements, type - are altered. Specifically, this function implements an algorithm to transform cross-validation dataset for individual plots. The function returns Transformed cross-validation dataset (a tibble).
}
\keyword{internal}
